Move float matchers to single file
As 4.0 dictated some breaking changes, this commit will move the FloatMatchers to a single file in the correct package.
